fix: extract comms conversation surface
